Static OffenceController
