+++ WPFWijkagent/WPFWijkagent/AddOffenceDialogue.xaml.cs
@@ -12,19 +12,14 @@ namespace WijkagentWPF
     {
         public OffenceCategories categories = new OffenceCategories();
 
-        private OffenceController Controller { get; }
-
         public Location Location { get; set; }
 
 
         //Create the AddOffenceDialogue. this method initializes all the components used by the AddOffenceDialogue
-        public AddOffenceDialogue(OffenceController controller)
+        public AddOffenceDialogue()
         {
             //Initializes itself (the Window)
             InitializeComponent();
-
-            //init controller and window so these properties can be used later on
-            Controller = controller;
 
             //add all enum categories to ComboBox so they can be selected
             InitializeCategories();
@@ -59,7 +54,7 @@ namespace WijkagentWPF
 
                 if (date != null && date < DateTime.Now && wpfCBCategorie.SelectedItem != null && Location != null)
                 {
-                    Controller.SetOffenceData(wpfTBOmschrijving.Text, (OffenceCategories)wpfCBCategorie.SelectedItem, date, Location);
+                    OffenceController.SetOffenceData(wpfTBOmschrijving.Text, (OffenceCategories)wpfCBCategorie.SelectedItem, date, Location);
                     this.Close();
                 }
             }
